push de front

+++ db/schema.rb
@@ -19,7 +19,7 @@ ActiveRecord::Schema.define(version: 2019_08_25_000727) do
     t.string "rfc", null: false
     t.string "ccyisocode", null: false
     t.string "place_generated", null: false
-    t.integer "subtotal", null: false
+    t.float "subtotal", null: false
     t.datetime "date", null: false
     t.boolean "vigente", null: false
     t.integer "quantity", null: false
